+++ lib/maven.sh
@@ -28,15 +28,52 @@ _mvn_cmd_opts() {
   fi
 }
 
+_mvn_settings_opt() {
+  local home="${1}"
+  local mavenInstallDir="${2}"
+
+  if [ -n "$MAVEN_SETTINGS_PATH" ]; then
+    mcount "mvn.settings.path"
+    echo -n "-s $MAVEN_SETTINGS_PATH"
+  elif [ -n "$MAVEN_SETTINGS_URL" ]; then
+    local settingsXml="${mavenInstallDir}/.m2/settings.xml"
+    mkdir -p $(dirname ${settingsXml})
+    curl --retry 3 --silent --max-time 10 --location $MAVEN_SETTINGS_URL --output ${settingsXml}
+    mcount "mvn.settings.url"
+    if [ -f ${settingsXml} ]; then
+      echo -n "-s ${settingsXml}"
+    else
+      mcount "mvn.settings.url.fail"
+      error "Could not download settings.xml from the URL defined in MAVEN_SETTINGS_URL!"
+      return 1
+    fi
+  elif [ -f ${home}/settings.xml ]; then
+    mcount "mvn.settings.file"
+    echo -n "-s ${home}/settings.xml"
+  else
+    mcount "mvn.settings.default"
+    echo -n ""
+  fi
+}
+
 has_maven_wrapper() {
   local home=${1}
   if [ -f $home/mvnw ] &&
       [ -f $home/.mvn/wrapper/maven-wrapper.jar ] &&
       [ -f $home/.mvn/wrapper/maven-wrapper.properties ] &&
-      [ -z "$(detect_maven_version $home)"]; then
+      [ -z "$(detect_maven_version $home)" ]; then
     return 0;
   else
     return 1;
+  fi
+}
+
+get_cache_status() {
+  local cacheDir=${1}
+  if [ ! -d ${cacheDir}/.m2 ]; then
+    echo "not-found"
+  else
+    echo "valid"
   fi
 }
 
@@ -45,35 +82,32 @@ run_mvn() {
   local home=${2}
   local mavenInstallDir=${3}
 
+  mkdir -p ${mavenInstallDir}
   if has_maven_wrapper $home; then
     cache_copy ".m2/wrapper" $mavenInstallDir $home
     chmod +x $home/mvnw
     local mavenExe="./mvnw"
+    mcount "mvn.version.wrapper"
   else
     cd $mavenInstallDir
+    let start=$(nowms)
     install_maven ${mavenInstallDir} ${home}
+    mtime "mvn.${scope}.time" "${start}"
     PATH="${mavenInstallDir}/.maven/bin:$PATH"
     local mavenExe="mvn"
     cd $home
   fi
 
-  if [ -n "$MAVEN_SETTINGS_PATH" ]; then
-    local mvn_settings_opt="-s $MAVEN_SETTINGS_PATH"
-  elif [ -n "$MAVEN_SETTINGS_URL" ]; then
-    status_pending "Installing settings.xml"
-    mkdir -p ${mavenInstallDir}/.m2
-    curl --retry 3 --silent --max-time 10 --location $MAVEN_SETTINGS_URL --output ${mavenInstallDir}/.m2/settings.xml
-    status_done
-    local mvn_settings_opt="-s $mavenInstallDir/.m2/settings.xml"
-  elif [ -f ${home}/settings.xml ]; then
-    local mvn_settings_opt="-s ${home}/settings.xml"
-  fi
+  local mvn_settings_opt="$(_mvn_settings_opt ${home} ${mavenInstallDir})"
 
   export MAVEN_OPTS="$(_mvn_java_opts ${scope} ${home} ${mavenInstallDir})"
 
   cd $home
   local mvnOpts="$(_mvn_cmd_opts ${scope})"
   status "Executing: ${mavenExe} ${mvnOpts}"
+
+  local cache_status="$(get_cache_status ${mavenInstallDir})"
+  let start=$(nowms)
   ${mavenExe} -DoutputFile=target/mvn-dependency-list.log -B ${mvn_settings_opt} ${mvnOpts} | indent
 
   if [ "${PIPESTATUS[*]}" != "0 0" ]; then
@@ -81,15 +115,17 @@ run_mvn() {
 We're sorry this build is failing! If you can't find the issue in application code,
 please send an email at support@scalingo.com -- The Scalingo Team"
   fi
+
+  mtime "mvn.${scope}.time" "${start}"
+  mtime "mvn.${scope}.time.cache.${cache_status}" "${start}"
 }
 
 write_mvn_profile() {
   local home=${1}
-  local mvnBinDir=${home}/.maven/bin
   mkdir -p ${home}/.profile.d
   cat << EOF > ${home}/.profile.d/maven.sh
-export M2_HOME="${home}/.maven"
-export MAVEN_OPTS="$(_mvn_java_opts "test" ${home} ${home})"
-export PATH="${mvnBinDir}:\$PATH"
+export M2_HOME="\$HOME/.maven"
+export MAVEN_OPTS="$(_mvn_java_opts "test" "\$HOME" "\$HOME")"
+export PATH="\$M2_HOME/bin:\$PATH"
 EOF
 }