+++ spec/spring_spec.rb
@@ -15,10 +15,9 @@ describe "Spring" do
         it "builds a war" do
           app.deploy do |app|
             sleep(10) # :(
-            expect(app.output).to include("Installing OpenJDK #{jdk_version}")
+            expect(app.output).to include("Installing JDK #{jdk_version}")
             expect(app.output).to match(%r{Building war: /tmp/.*/target/spring-boot-example-1.0-SNAPSHOT.war})
             expect(app.output).not_to match(%r{Building jar: /tmp/.*/target/spring-boot-example-1.0-SNAPSHOT.jar})
-            expect(app.output).not_to include("Installing settings.xml")
             expect(app.output).not_to include("BUILD FAILURE")
 
             expect(successful_body(app)).to include("Create a New Appointment")
@@ -32,11 +31,10 @@ describe "Spring" do
         it "builds an executable jar" do
           app.deploy do |app|
             sleep(10) # :(
-            expect(app.output).to include("Installing OpenJDK #{jdk_version}")
+            expect(app.output).to include("Installing JDK #{jdk_version}")
             expect(app.output).not_to include("Installing Maven")
             expect(app.output).not_to match(%r{Building war: /tmp/.*/target/spring-boot-example-1.0-SNAPSHOT.war})
             expect(app.output).to match(%r{Building jar: /tmp/.*/target/spring-boot-example-1.0-SNAPSHOT.jar})
-            expect(app.output).not_to include("Installing settings.xml")
             expect(app.output).not_to include("BUILD FAILURE")
 
             expect(successful_body(app)).to include("Create a New Appointment")
@@ -56,10 +54,9 @@ describe "Spring" do
           it "creates a default process type" do
             app.deploy do |app|
               sleep(10) # :(
-              expect(app.output).to include("Installing OpenJDK #{jdk_version}")
+              expect(app.output).to include("Installing JDK #{jdk_version}")
               expect(app.output).not_to match(%r{Building war: /tmp/.*/target/spring-boot-example-1.0-SNAPSHOT.war})
               expect(app.output).to match(%r{Building jar: /tmp/.*/target/spring-boot-example-1.0-SNAPSHOT.jar})
-              expect(app.output).not_to include("Installing settings.xml")
               expect(app.output).not_to include("BUILD FAILURE")
               expect(app.output).to include("Procfile declares types     -> (none)")
               expect(app.output).to include("Default types for buildpack -> web")